requirements.txt
# Core ML Libraries
# torch>=2.0.0
# torchvision>=0.15.0
timm>=0.9.0

# Tabular Models
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0  # Fixed: NumPy 2.x incompatible with PyTorch 1.13.1 and older packages
rasterio>=1.3.0
geopandas>=0.13.0
Pillow>=10.0.0

# Utilities
tqdm==4.67.1
pyyaml==6.0.3
matplotlib==3.10.8
seaborn==0.13.2
scipy==1.15.3  # Scientific computing
requests==2.28.1  # HTTP library
joblib==1.5.3  # Parallel processing
beautifultable==1.1.0  # Table formatting (used in logging)
einops==0.8.1  # Tensor operations
opencv-python-headless==4.12.0.88  # Image processing (headless version)

# Hyperparameter Optimization
optuna>=3.3.0

# Ensemble
mlxtend>=0.22.0

# Other
pyproj>=3.6.0
shapely>=2.0.0

# TIP (Tabular-Image Pre-training) Dependencies
# Compatible versions: PyTorch 1.13.1 + PyTorch Lightning 1.7.7 + lightning-bolts 0.5.0
hydra-core==1.3.2
omegaconf==2.3.0
pytorch-lightning==1.7.7
torchmetrics<0.10.0  # Compatible with PyTorch 1.13.1
lightning-bolts==0.5.0
lightning-utilities==0.15.2  # PyTorch Lightning utilities
albumentations==2.0.8
wandb==0.23.1
tensorboard==2.20.0
lightly==1.5.22  # Self-supervised learning library (used in TIP)

